% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bootStrap}
\alias{bootStrap}
\title{bootStrap}
\usage{
bootStrap(
  ...,
  feature_file = system.file("extdata", "tss_locations.txt", package = "svBreaks"),
  feature = "tss",
  n = 10
)
}
\description{
Plot reav vs sim data bootstrapped n times
}
