% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneEnrichment.R
\name{bpAllGenes}
\alias{bpAllGenes}
\title{Calculate the enrichment of genes affected by SVs (as opposed to breakpoint hits)}
\usage{
bpAllGenes(gene_lengths_in = system.file("extdata", "gene_lengths.txt",
  package = "svBreaks"), affected_genes = system.file("extdata",
  "all_genes_filtered.txt", package = "svBreaks"), n = 3,
  genome_length = 118274340)
}
\description{
Calculate the enrichment of genes affected by SVs (as opposed to breakpoint hits)
}
\keyword{enrichment}
