% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureEnrichment.R
\name{bpFeatureEnrichmentPlot}
\alias{bpFeatureEnrichmentPlot}
\title{Plot the enrichment of SVs in genomic features}
\usage{
bpFeatureEnrichmentPlot(...)
}
\description{
Plot the enrichment of SVs in genomic features
}
\keyword{enrichment}
