% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{bpFeatures}
\alias{bpFeatures}
\title{bpFeatures
Get some basic stats for breakpoints}
\usage{
bpFeatures(notch = 0)
}
\arguments{
\item{notch}{- Set to 1 to exclude Notch events}
}
\description{
bpFeatures
Get some basic stats for breakpoints
}
\keyword{stats}
