% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneEnrichment.R
\name{bpGeneEnrichment}
\alias{bpGeneEnrichment}
\title{Calculate the enrichment of SVs in genes}
\usage{
bpGeneEnrichment(gene_lengths = "data/gene_lengths.txt", n = 3,
  genome_length = 118274340, print = NA)
}
\description{
Calculate the enrichment of SVs in genes
}
\keyword{enrichment}
