% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneEnrichment.R
\name{bpGeneEnrichmentPlot}
\alias{bpGeneEnrichmentPlot}
\title{Plot the enrichment of SVs in genes}
\usage{
bpGeneEnrichmentPlot(n = 2)
}
\description{
Plot the enrichment of SVs in genes
}
\keyword{enrichment}
