% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpRegioner.R
\name{bpRegioneR}
\alias{bpRegioneR}
\title{Run regioneR for to regions in bed format}
\usage{
bpRegioneR(
  ...,
 
    regionA = "~/Desktop/misc_bed/breakpoints/Notch_CFS/notch_10kb_merged_breakpoints.bed",
  regionB = "~/GitHub/BardinLab/meme/out/notch_CFS/fimo_out/motif2_merged.bed",
 
    mappable_regions = "~/Documents/Curie/Data/Genomes/Dmel_v6.12/Mappability/dmel6_mappable.bed",
  chrom_lengths = "~/Documents/Curie/Data/Genomes/Dmel_v6.12/chrom.sizes.txt",
  slop,
  limit2chrom = TRUE,
  chrom = "X",
  from,
  to,
  n = 100,
  plot = TRUE,
  bins = "auto",
  region = NULL
)
}
\arguments{
\item{regionA}{RegionA [Required]}

\item{regionB}{RegionB [Required]}

\item{mappable_regions}{Bed file with mappable geneome [Required]}

\item{chrom_lengths}{Lengths for for chromosomes}

\item{slop}{Add this much slop to each region}

\item{limit2chrom}{Restrict permutaions to one chromsome [Default: TRUE]}

\item{chrom}{Chromosome to run permutations on if limit2chrom=TRUE [Default: "X"]}

\item{from}{Restrict to region between from and to}

\item{to}{Restrict to region between from and to}

\item{n}{Number of permutations to run}

\item{plot}{Return plot or not? [Default: TRUE]}

\item{region}{Limit to region (chr:start-stop) [Default: NULL]}
}
\description{
Run regioneR for to regions in bed format
}
