% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2motif.R
\name{bpSim}
\alias{bpSim}
\title{bpSim}
\usage{
bpSim(intervals = system.file("extdata", "intervals.bed", package =
  "svBreaks"), N = 1000, write = F)
}
\arguments{
\item{intervals}{File containing genomic regions within which to simulate SNVs [Default 'data/intervals.bed]}

\item{N}{Number of random breakpoints to generate [Default nrow(bp_data)]}
}
\description{
Generate simulated SV breakpoints acroos genomic regions (e.g. mappable regions)
}
\keyword{sim}
