% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionEnrichment.R
\name{delsRegionEnrichment}
\alias{delsRegionEnrichment}
\title{asdas}
\usage{
delsRegionEnrichment(bedFiles = "bed", region = TRUE, breakpoints = NA,
  genome_length = 118274340)
}
\description{
asdas
}
\keyword{overlap}
\keyword{regions}
