% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2motif.R
\name{dist2Motif}
\alias{dist2Motif}
\title{dist2Motif
Calculate the distance from each breakpoint to closest motif}
\usage{
dist2Motif(
  ...,
  breakpoints = NA,
  feature_file = system.file("extdata", "tss_locations.txt", package = "svBreaks"),
  sim = NA,
  print = 0,
  send = 0,
  feature = "tss",
  keep = NULL,
  position = "centre"
)
}
\description{
dist2Motif
Calculate the distance from each breakpoint to closest motif
}
\keyword{motif}
