% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2motif.R
\name{distOverlay}
\alias{distOverlay}
\title{Calculate the distance from each breakpoint to closest motif
Overlay the same number of random simulated breakpoints}
\usage{
distOverlay(feature_file = system.file("extdata", "tss_locations.txt", package
  = "svBreaks"), feature = "tss", lim = 10, byChrom = NA, n = 10)
}
\description{
Calculate the distance from each breakpoint to closest motif
Overlay the same number of random simulated breakpoints
}
\keyword{motif}
