% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2motif.R
\name{dtableShuffle}
\alias{dtableShuffle}
\title{dtableShuffle}
\usage{
dtableShuffle(nSites = 1000, byChrom = NA)
}
\arguments{
\item{intervals}{File containing genomic regions within which to simulate SNVs [Default 'data/intervals.bed]}

\item{N}{Number of random breakpoints to generate [Default nrow(bp_data)]}
}
\description{
Generate simulated SV breakpoints acroos genomic regions (e.g. mappable regions)
}
\keyword{sim}
