% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notch.R
\name{geneHit}
\alias{geneHit}
\title{Print barplot showing samples with SVs affecting specified gene}
\usage{
geneHit(..., all_samples, filter_gene = "N", plot = TRUE)
}
\arguments{
\item{filter_gene}{The gene of interest [Default: N]}

\item{plot}{Show barplot [Default: TRUE]}

\item{infile}{File to process [Required]}
}
\description{
Print barplot showing samples with SVs affecting specified gene
}
