% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2motif.R
\name{generateData}
\alias{generateData}
\title{generateData
Prepare data for dist2motif}
\usage{
generateData(..., breakpoints = NA, sim = NA)
}
\description{
generateData
Prepare data for dist2motif
}
\keyword{simulate}
