% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2motif2.R
\name{generateData2}
\alias{generateData2}
\title{generateData2
Prepare data for dist2motif}
\usage{
generateData2(..., df, breakpoints, sim = FALSE, chroms = c("2L", "2R",
  "3L", "3R", "4", "X", "Y"))
}
\description{
generateData2
Prepare data for dist2motif
}
\keyword{simulate}
