% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{genesbySample}
\alias{genesbySample}
\title{genesbySample
Plot the number of structural breakpoints per sample}
\usage{
genesbySample(...,
  affected_genes = "inst/extdata/all_genes_filtered.txt",
  genome_length = 118274340)
}
\description{
genesbySample
Plot the number of structural breakpoints per sample
}
\keyword{genes}
