% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeDist.R
\name{genotypeTypeCount}
\alias{genotypeTypeCount}
\title{genotypeTypeCount
Plot the contribution of sv tpyes to toal load per genotype}
\usage{
genotypeTypeCount(...)
}
\description{
genotypeTypeCount
Plot the contribution of sv tpyes to toal load per genotype
}
\keyword{genotype}
