% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(
  ...,
  infile = system.file("extdata", "all_bps_merged.txt", package = "svBreaks"),
  attach_info = system.file("extdata", "samples_names_conversion.txt", package =
    "svBreaks"),
  gene_lengths_file = system.file("extdata", "gene_lengths.txt", package = "svBreaks"),
  expression_data = system.file("extdata", "isc_genes_rnaSeq.txt", package =
    "svBreaks"),
  expression_source = "flygut"
)
}
\arguments{
\item{infile}{File to process [Required]}
}
\value{
Dataframe
}
\description{
Function to clean cnv files
}
\keyword{breakpoints}
\keyword{data}
\keyword{parse}
