% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getMissingSamples}
\alias{getMissingSamples}
\title{get sample names}
\usage{
getMissingSamples(
  ...,
  df = NULL,
 
    all_samples_info = "~/Desktop/script_test/mutationProfiles/data/samples_names_conversion.txt"
)
}
\description{
get sample names
}
