% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionEnrichment.R
\name{ggVolcano}
\alias{ggVolcano}
\title{Plot the enrichment of SVs in genomic features}
\usage{
ggVolcano(..., df)
}
\description{
Plot the enrichment of SVs in genomic features
}
\keyword{enrichment}
