% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanims.R
\name{mechansimSize}
\alias{mechansimSize}
\title{Plot mechanims vs length
mechansimSize}
\usage{
mechansimSize(..., bp_data = NULL, plot = TRUE)
}
\arguments{
\item{infile}{File to process [Required]}
}
\description{
Function to plot the sizes SVs annotated as being underscored by mechansims
}
\keyword{parse}
