% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionEnrichment.R
\name{mergeOverlaps}
\alias{mergeOverlaps}
\title{Implements bedtools merge to merge overlapping regions in bedfile}
\usage{
mergeOverlaps(f, dataframe = FALSE)
}
\description{
Implements bedtools merge to merge overlapping regions in bedfile
}
\keyword{bed}
\keyword{merge}
