% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notch.R
\name{notchDels}
\alias{notchDels}
\title{notchDels
Plot the size of deletions affecting Notch}
\usage{
notchDels(infile = "inst/extdata/Notch_hits.txt")
}
\description{
notchDels
Plot the size of deletions affecting Notch
}
\keyword{Notch}
