% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notch.R
\name{notchFilt}
\alias{notchFilt}
\title{Functions related to Notch
notchFilt}
\usage{
notchFilt(..., keep = NULL, start = 2700000, stop = 3400000)
}
\arguments{
\item{infile}{File to process [Required]}
}
\value{
Dataframe
}
\description{
Function to filter in/out events affecting a locus (in this case the Drosophila Notch locus)
}
\keyword{parse}
