% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notch.R
\name{notchHits}
\alias{notchHits}
\title{notchHits
Plot the breakpint density over Notch}
\usage{
notchHits(..., all_samples, drivers = c("N", "kuz"),
  show_samples = FALSE, bp_density = FALSE, from = 2.7, to = 3.4,
  ticks = 50)
}
\description{
notchHits
Plot the breakpint density over Notch
}
\keyword{Notch}
