% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2motif.R
\name{plotdistanceOverlay}
\alias{plotdistanceOverlay}
\title{plotdistanceOverlay}
\usage{
plotdistanceOverlay(
  ...,
  d,
  from = "bps",
  feature = "tss",
  lim = 10,
  byChrom = NA,
  n = 10,
  write = TRUE,
  facetPlot = TRUE
)
}
\arguments{
\item{d}{Dataframe containing combined real + sim data (d <- distOverlay())}
}
\description{
Plot the distance overlay
}
\keyword{distance}
