% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2motif2.R
\name{plotdistanceOverlay2}
\alias{plotdistanceOverlay2}
\title{plotdistanceOverlay}
\usage{
plotdistanceOverlay2(..., distances, from = "bps", lim = 2.5, n = 2,
  position = "centre", histo = FALSE, binWidth = 500)
}
\arguments{
\item{d}{Dataframe containing combined real + sim data (d <- distOverlay())}
}
\description{
Plot the distance overlay
}
\keyword{distance}
