% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionEnrichment.R
\name{plotMonteCarlo}
\alias{plotMonteCarlo}
\title{Plot the result of a monte carlo simulation (n shuffles of feature y)}
\usage{
plotMonteCarlo(x)
}
\description{
Plot the result of a monte carlo simulation (n shuffles of feature y)
}
\keyword{expected}
\keyword{frequency}
