% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setCols}
\alias{setCols}
\title{Set colour scheme for manual colouring}
\usage{
setCols(df, col, fill = "Y", set = "Pastel2")
}
\value{
colourscale object
}
\description{
Set colour scheme for manual colouring
}
\keyword{colour}
\keyword{scheme}
