% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{showSamples}
\alias{showSamples}
\title{showSamples}
\usage{

  showSamples(infile = "/Users/Nick_curie/Desktop/parserTest/filtered_231018/summary/merged/all_bps_filtered.txt")
}
\arguments{
\item{infile}{File to process [Required]}
}
\description{
A helper function to print the sample names
}
\keyword{samples}
