% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2motif.R
\name{simSig}
\alias{simSig}
\title{simSig}
\usage{
simSig(r, s, test = NA, max_dist = 5000)
}
\arguments{
\item{real_data}{Dataframe containing real data (as produced by generateData())}
}
\description{
Calculate the median distance between breakpoints (per-chrom) and feature
and perform test
}
\keyword{sim}
