% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeDist.R
\name{sizeDist}
\alias{sizeDist}
\title{A function to plot the size distribution of SV types
sizeDist
Plot the size distribution for different types of structual variants}
\usage{
sizeDist()
}
\description{
A function to plot the size distribution of SV types
sizeDist
Plot the size distribution for different types of structual variants
}
\keyword{size}
