% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionEnrichment.R
\name{subtractUnmappable}
\alias{subtractUnmappable}
\title{Implements bedtools subtract to remove unmappable regions in bedfiles}
\usage{
subtractUnmappable(f,
  u = "~/Documents/Curie/Data/Genomes/Dmel_v6.12/Mappability/dmel6_unmappable_75.bed",
  dataframe = FALSE)
}
\description{
Implements bedtools subtract to remove unmappable regions in bedfiles
}
\keyword{bed}
\keyword{subtract}
