% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{svsbySample}
\alias{svsbySample}
\title{svsbySample
Plot the number of structural breakpoints per sample}
\usage{
svsbySample(..., bp_data = NULL, colSample = NA)
}
\description{
svsbySample
Plot the number of structural breakpoints per sample
}
\keyword{samples}
