% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{svTypes}
\alias{svTypes}
\title{svTypes}
\usage{
svTypes(
  ...,
  bp_data = NULL,
  title = expression("Structural variants per sample (" ~ italic("Notch") ~
    "excluded )"),
  object = "type",
  keep = FALSE,
  plot = T
)
}
\description{
Plot counts of different svTypes
}
