% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{swapSampleNames}
\alias{swapSampleNames}
\title{Swap sample names}
\usage{
swapSampleNames(
  df,
 
    attach_info = "~/Desktop/script_test/mutationProfiles/data/samples_names_conversion.txt"
)
}
\description{
Swap sample names
}
