% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notch.R
\name{tally_hits}
\alias{tally_hits}
\title{Tally type of sv events in specified gene}
\usage{
tally_hits(
  ...,
  all_samples,
  bp_data,
  filter_gene = "N",
  plot = FALSE,
  freq = FALSE
)
}
\arguments{
\item{filter_gene}{The gene of interest [Default: N]}

\item{infile}{File to process [Required]}
}
\description{
Tally type of sv events in specified gene
}
