% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionEnrichment.R
\name{Volcano}
\alias{Volcano}
\title{Plot the enrichment of SVs in genomic features}
\usage{
Volcano(d, byq = FALSE)
}
\description{
Plot the enrichment of SVs in genomic features
}
\keyword{enrichment}
