% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionEnrichment.R
\name{writeBed}
\alias{writeBed}
\title{Simple function to write out bedfiels - useful for checking after merging/subtracting...}
\usage{
writeBed(df, outDir = getwd(), name = "regions.bed",
  svBreaks = FALSE)
}
\description{
Simple function to write out bedfiels - useful for checking after merging/subtracting...
}
\keyword{bed}
