% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{bpRainfall}
\alias{bpRainfall}
\title{bpRainfall
Plot log10 distances between snvs as rainfall plot}
\usage{
bpRainfall()
}
\description{
bpRainfall
Plot log10 distances between snvs as rainfall plot
}
\keyword{rainfall}
