% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionEnrichment.R
\name{bpRegionEnrichment}
\alias{bpRegionEnrichment}
\title{Calculate the enrichment of SVs in genomic regions}
\usage{
bpRegionEnrichment(bedDir = "inst/extdata/bed/mappable", breakpoints = NA,
  slop = 0, plot = TRUE, genome_length = 118274340, intersect = FALSE,
  write = FALSE)
}
\description{
Calculate the enrichment of SVs in genomic regions
}
\keyword{enrichment}
