% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionEnrichment.R
\name{bpRegionEnrichmentPlot}
\alias{bpRegionEnrichmentPlot}
\title{Plot the enrichment of SVs in genomic features}
\usage{
bpRegionEnrichmentPlot(...)
}
\description{
Plot the enrichment of SVs in genomic features
}
\keyword{enrichment}
