% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{bpStats}
\alias{bpStats}
\title{bpStats
Get some basic stats for breakpoints}
\usage{
bpStats(colSample = NA)
}
\arguments{
\item{colSample}{Set to <samplename> to colour that sample red in the plot}
}
\description{
bpStats
Get some basic stats for breakpoints
}
\keyword{stats}
