% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2motif2.R
\name{dist2motif2}
\alias{dist2motif2}
\title{dist2Motif2
Calculate the distance from each breakpoint to closest motif in a directory of files}
\usage{
dist2motif2(
  ...,
  df,
  breakpoints,
  feature_file,
  featureDir = system.file("extdata", "features", package = "svBreaks"),
  chroms = c("2L", "2R", "3L", "3R", "4", "X", "Y"),
  sim = FALSE,
  position = "centre"
)
}
\description{
dist2Motif2
Calculate the distance from each breakpoint to closest motif in a directory of files
}
\keyword{motif}
