% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2motif2.R
\name{distOverlay2}
\alias{distOverlay2}
\title{Calculate the distance from each breakpoint to closest motif
Overlay the same number of random simulated breakpoints}
\usage{
distOverlay2(..., df, breakpoints, feature_file,
  featureDir = system.file("extdata", "features", package = "svBreaks"),
  from = "bps", chroms = c("2L", "2R", "3L", "3R", "4", "X", "Y"),
  lim = 2.5, n = 2, plot = TRUE, position = "centre",
  in_window = 5, write = FALSE, out_dir)
}
\description{
Calculate the distance from each breakpoint to closest motif
Overlay the same number of random simulated breakpoints
}
\keyword{motif}
