% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notch.R
\name{geneHit}
\alias{geneHit}
\title{Print barplot showing samples with SVs affecting specified gene}
\usage{
geneHit(..., all_samples, drivers = c("N"), plot = TRUE,
  show_sample = TRUE)
}
\arguments{
\item{plot}{Show barplot [Default: TRUE]}

\item{infile}{File to process [Required]}

\item{filter_gene}{The gene of interest [Default: N]}
}
\description{
Print barplot showing samples with SVs affecting specified gene
}
