% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notch.R
\name{geneHit}
\alias{geneHit}
\title{Print barplot showing samples with SVs affecting specified gene}
\usage{
geneHit(...,
  all_samples = "/Users/Nick_curie/Desktop/parserTest/filtered_231018/summary/merged/all_samples.txt",
  filter_gene = "N")
}
\arguments{
\item{infile}{File to process [Required]}
}
\description{
Print barplot showing samples with SVs affecting specified gene
}
