% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(..., infile = "inst/extdata/all_bps_filtered.txt",
  gene_lengths_file = system.file("extdata", "gene_lengths.txt", package =
  "svBreaks"), expression_data = system.file("extdata",
  "isc_genes_rnaSeq.csv", package = "svBreaks"))
}
\arguments{
\item{infile}{File to process [Required]}
}
\value{
Dataframe
}
\description{
Function to clean cnv files
}
\keyword{breakpoints}
\keyword{data}
\keyword{parse}
