% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(...,
  infile = "/Users/Nick_curie/Desktop/parserTest/filtered_231018/summary/merged/all_bps_mech.txt",
  attach_info = "../mutationProfiles/data/samples_names_conversion.txt",
  gene_lengths_file = system.file("extdata", "gene_lengths.txt", package
  = "svBreaks"), expression_source = "flygut",
  expression_data = system.file("extdata", "isc_genes_rnaSeq.txt",
  package = "svBreaks"), exclude = TRUE)
}
\arguments{
\item{infile}{File to process [Required]}
}
\value{
Dataframe
}
\description{
Function to clean cnv files
}
\keyword{breakpoints}
\keyword{data}
\keyword{parse}
