% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanims.R
\name{mechansimSize}
\alias{mechansimSize}
\title{Plot mechanims vs length
mechansimSize}
\usage{
mechansimSize(
  ...,
  bp_data = NULL,
  infile = "~/Desktop/parserTest/filtered_231018/summary/merged/all_bps_mech.txt",
  plot = TRUE
)
}
\arguments{
\item{infile}{File to process [Required]}
}
\description{
Function to plot the sizes SVs annotated as being underscored by mechansims
}
\keyword{parse}
