% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{notchFilt}
\alias{notchFilt}
\title{notchFilt}
\usage{
notchFilt(keep = 0)
}
\arguments{
\item{infile}{File to process [Required]}
}
\value{
Dataframe
}
\description{
Function to filter in/out events affecting a locus (in this case the Drosophila Notch locus)
}
\keyword{parse}
