% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notch.R
\name{notchHits}
\alias{notchHits}
\title{notchHits
Plot the breakpint density over Notch}
\usage{
notchHits(
  ...,
  all_samples,
  all_samples_df = NULL,
  drivers = c("N"),
  show_samples = FALSE,
  barchart = FALSE,
  bp_density = FALSE,
  adjust = 0.3,
  from = 2.7,
  to = 3.5,
  ticks = 50
)
}
\description{
notchHits
Plot the breakpint density over Notch
}
\keyword{Notch}
