% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notch.R
\name{notchHits}
\alias{notchHits}
\title{notchHits
Plot the breakpint density over Notch}
\usage{
notchHits(infile = "data/Notch_hits.txt")
}
\description{
notchHits
Plot the breakpint density over Notch
}
\keyword{Notch}
