% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionEnrichment.R
\name{plotMonteCarlo}
\alias{plotMonteCarlo}
\title{Plot the result of a monte carlo simulation (n shuffles of feature y)
Expects dataframe returned from bpRegionEnrichment()}
\usage{
plotMonteCarlo(x, bindWith = 10)
}
\description{
Plot the result of a monte carlo simulation (n shuffles of feature y)
Expects dataframe returned from bpRegionEnrichment()
}
\keyword{expected}
\keyword{frequency}
