% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{showSamples}
\alias{showSamples}
\title{showSamples}
\usage{
showSamples(..., bp_data = NULL)
}
\arguments{
\item{infile}{File to process [Required]}
}
\description{
A helper function to print the sample names
}
\keyword{samples}
