% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeDist.R
\name{sizeDist}
\alias{sizeDist}
\title{sizeDist
Plot the size distribution for different types of structual variants}
\usage{
sizeDist(...)
}
\description{
sizeDist
Plot the size distribution for different types of structual variants
}
\keyword{size}
